Moved stubbing code into its own file
